rename tally c files
