+++ xprof/btx_tally.c
@@ -1,5 +1,5 @@
 #include <babeltrace2/babeltrace.h>
-#include "tally.h"
+#include "btx_tally.h"
 /* Mandatory */
 BT_PLUGIN_MODULE();
